Share common init code, lint exceptions cleanup
