+++ src/extension.js
@@ -15,16 +15,16 @@ import {
     gettext as _,
     ngettext
 } from 'resource:///org/gnome/shell/extensions/extension.js';
-import * as Utils from './shell/utils.js';
 
-// eslint-disable-next-line no-redeclare
 import Config from './config.mjs';
 import * as Clipboard from './shell/clipboard.js';
 import * as Device from './shell/device.js';
 import * as Keybindings from './shell/keybindings.js';
 import * as Notification from './shell/notification.js';
 import * as Input from './shell/input.js';
+import * as Utils from './shell/utils.js';
 import Remote from './utils/remote.mjs';
+import setup from './utils/setup.mjs';
 
 const QuickSettingsMenu = Main.panel.statusArea.quickSettings;
 
@@ -358,7 +358,7 @@ export default class GSConnectExtension extends Extension {
 
     constructor(metadata) {
         super(metadata);
-        Utils.setupExtensionData(this.path);
+        setup(this.path);
 
         // If installed as a user extension, this checks the permissions
         // on certain critical files in the extension directory
