+++ src/gsconnect-preferences
@@ -40,38 +40,13 @@ function get_datadir() {
 imports.searchPath.unshift(get_datadir());
 imports.config.PACKAGE_DATADIR = imports.searchPath[0];
 
-// FIXME when ESM: This block of code is duplicate to setupExtensionData() in src/shell/utils.js
-// Ensure config.js is setup properly
-const Config = imports.config;
-const userDir = GLib.build_filenamev([GLib.get_user_data_dir(), 'gnome-shell']);
-
-if (Config.PACKAGE_DATADIR.startsWith(userDir)) {
-    Config.IS_USER = true;
-
-    Config.GSETTINGS_SCHEMA_DIR = `${Config.PACKAGE_DATADIR}/schemas`;
-    Config.PACKAGE_LOCALEDIR = `${Config.PACKAGE_DATADIR}/locale`;
-}
-
-// Init GResources
-Gio.Resource.load(
-    GLib.build_filenamev([Config.PACKAGE_DATADIR, `${Config.APP_ID}.gresource`])
-)._register();
-
-// Init GSchema
-Config.GSCHEMA = Gio.SettingsSchemaSource.new_from_directory(
-    Config.GSETTINGS_SCHEMA_DIR,
-    Gio.SettingsSchemaSource.get_default(),
-    false
-);
-
-// Init Gettext
-String.prototype.format = imports.format.format;
-imports.gettext.bindtextdomain(Config.APP_ID, Config.PACKAGE_LOCALEDIR);
-globalThis._ = GLib.dgettext.bind(null, Config.APP_ID);
-globalThis.ngettext = GLib.dngettext.bind(null, Config.APP_ID);
-
+// Bootstrap
+const {setup, setupGettext} = imports.utils.setup;
+setup(imports.config.PACKAGE_DATADIR);
+setupGettext();
 
 // Local Imports
+const Config = imports.config;
 const Settings = imports.preferences.service;
