+++ src/service/utils/setup.js
@@ -4,13 +4,12 @@
 
 'use strict';
 
-const Gettext = imports.gettext;
-
 const Gio = imports.gi.Gio;
 const GIRepository = imports.gi.GIRepository;
 const GLib = imports.gi.GLib;
 
 const Config = imports.config;
+const {setup, setupGettext} = imports.utils.setup;
 
 
 // Promise Wrappers
@@ -53,22 +52,16 @@ Gio._promisify(Gio.TlsConnection.prototype, 'handshake_async');
 Gio._promisify(Gio.DtlsConnection.prototype, 'handshake_async');
 
 
-// FIXME when ESM: Shares code with src/shell/utils.js, gsconnect-preferences for gettext initialisation too
 // User Directories
 Config.CACHEDIR = GLib.build_filenamev([GLib.get_user_cache_dir(), 'gsconnect']);
 Config.CONFIGDIR = GLib.build_filenamev([GLib.get_user_config_dir(), 'gsconnect']);
 Config.RUNTIMEDIR = GLib.build_filenamev([GLib.get_user_runtime_dir(), 'gsconnect']);
 
+// Bootstrap
+setup(Config.PACKAGE_DATADIR);
+setupGettext();
 
-// Ensure config.js is setup properly
-const userDir = GLib.build_filenamev([GLib.get_user_data_dir(), 'gnome-shell']);
-
-if (Config.PACKAGE_DATADIR.startsWith(userDir)) {
-    Config.IS_USER = true;
-
-    Config.GSETTINGS_SCHEMA_DIR = `${Config.PACKAGE_DATADIR}/schemas`;
-    Config.PACKAGE_LOCALEDIR = `${Config.PACKAGE_DATADIR}/locale`;
-
+if (Config.IS_USER) {
     // Infer libdir by assuming gnome-shell shares a common prefix with gjs;
     // assume the parent directory if it's not there
     let libdir = GIRepository.Repository.get_search_path().find(path => {
@@ -84,27 +77,6 @@ if (Config.PACKAGE_DATADIR.startsWith(userDir)) {
 
     Config.GNOME_SHELL_LIBDIR = libdir;
 }
-
-
-// Init Gettext
-String.prototype.format = imports.format.format;
-Gettext.bindtextdomain(Config.APP_ID, Config.PACKAGE_LOCALEDIR);
-globalThis._ = GLib.dgettext.bind(null, Config.APP_ID);
-globalThis.ngettext = GLib.dngettext.bind(null, Config.APP_ID);
-
-
-// Init GResources
-Gio.Resource.load(
-    GLib.build_filenamev([Config.PACKAGE_DATADIR, `${Config.APP_ID}.gresource`])
-)._register();
-
-
-// Init GSchema
-Config.GSCHEMA = Gio.SettingsSchemaSource.new_from_directory(
-    Config.GSETTINGS_SCHEMA_DIR,
-    Gio.SettingsSchemaSource.get_default(),
-    false
-);
 
 
 // Load DBus interfaces
