+++ src/shell/device.js
@@ -12,7 +12,6 @@ import * as PopupMenu from 'resource:///org/gnome/shell/ui/popupMenu.js';
 import {gettext as _} from 'resource:///org/gnome/shell/extensions/extension.js';
 import {getIcon} from './utils.js';
 
-// eslint-disable-next-line no-redeclare
 import * as GMenu from './gmenu.js';
 import Tooltip from './tooltip.js';
