+++ src/shell/utils.js
@@ -15,36 +15,6 @@ try {
 
 
 /**
- * Initialise and setup Config, GResources and GSchema.
- * @param {string} extensionPath - The absolute path to the extension directory
- */
-export function setupExtensionData(extensionPath) {
-    // Ensure config.js is setup properly
-    Config.PACKAGE_DATADIR = extensionPath;
-    const userDir = GLib.build_filenamev([GLib.get_user_data_dir(), 'gnome-shell']);
-
-    if (Config.PACKAGE_DATADIR.startsWith(userDir)) {
-        Config.IS_USER = true;
-
-        Config.GSETTINGS_SCHEMA_DIR = `${Config.PACKAGE_DATADIR}/schemas`;
-        Config.PACKAGE_LOCALEDIR = `${Config.PACKAGE_DATADIR}/locale`;
-    }
-
-    // Init GResources
-    Gio.Resource.load(
-        GLib.build_filenamev([Config.PACKAGE_DATADIR, `${Config.APP_ID}.gresource`])
-    )._register();
-
-    // Init GSchema
-    Config.GSCHEMA = Gio.SettingsSchemaSource.new_from_directory(
-        Config.GSETTINGS_SCHEMA_DIR,
-        Gio.SettingsSchemaSource.get_default(),
-        false
-    );
-}
-
-
-/**
  * Get a themed icon, using fallbacks from GSConnect's GResource when necessary.
  *
  * @param {string} name - A themed icon name
